backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
django-cors-headers>=4.0,<5.0
gunicorn>=21.0,<23.0